Add a Makefile to install everything in a proper location
Update all the scripts and configuration files to use the new locations.

This Makefile will be used by packaging tools to install the files
properly too. Thus this makes it much more easy for Kali to create a
proper package of the MANA toolkit.

I used "mana-toolkit" everywhere as "mana" is another software that does
already exist in Debian and Kali (it's a game).