Don't rely on system-wide setup of stunnel but use our own
The problem with installing files in /etc/stunnel/ is that those tunnels
might be auto-started indepently of mana and might conflict with other
usage of port 80. So we can start and stop stunnel4 ourselves and pass
it a configuration file stored in mana's configuration directory.

Use this opportunity to clean up the fact that we had two stunnel.conf
lying around.